+++ test/lastfm_archive/livebook_test.exs
@@ -5,6 +5,7 @@ defmodule LastfmArchive.LivebookTest do
 
   alias LastfmArchive.Behaviour.LastfmClient
   alias LastfmArchive.Cache
+  alias LastfmArchive.Cache.Server, as: CacheServer
   alias LastfmArchive.Livebook, as: LFM_LB
   alias LastfmArchive.Utils
 
@@ -14,7 +15,7 @@ defmodule LastfmArchive.LivebookTest do
   setup :verify_on_exit!
 
   setup do
-    {:ok, _pid} = start_supervised(%{id: @cache, start: {Cache, :start_link, [[name: @cache]]}})
+    {:ok, _pid} = start_supervised(%{id: @cache, start: {CacheServer, :start_link, [[name: @cache]]}})
     cache = %{{"a_user", 2006} => %{{1_138_752_000, 1_138_838_399} => {33, [:ok]}}}
 
     cache_file = "#{Utils.user_dir("a_user")}/#{Cache.cache_dir()}/2006"